ai-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-limiter==3.7.0
groq==0.9.0
chromadb==0.5.3
sentence-transformers==3.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
